Removed las common includes
