--- include/common/device/other/pin.hpp
@@ -1,42 +1,0 @@
-/*
-    Aversive++                                                           
-    Copyright (C) 2014 Eirbot                                            
-                                                                         
-    This program is free software: you can redistribute it and/or modify 
-    it under the terms of the GNU General Public License as published by 
-    the Free Software Foundation, either version 3 of the License, or    
-    (at your option) any later version.                                  
-                                                                         
-    This program is distributed in the hope that it will be useful,      
-    but WITHOUT ANY WARRANTY; without even the implied warranty of       
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        
-    GNU General Public License for more details.                         
-                                                                         
-    You should have received a copy of the GNU General Public License    
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef PIN_HPP
-#define PIN_HPP
-
-#include <device/input.hpp>
-#include <device/output.hpp>
-#include <device/device.hpp>
-
-enum class PinMode {
-  INPUT,
-  OUTPUT
-};
-
-template<int ID>
-class Pin : public Device, public Input<bool>, public Output<bool> {
-public:
-  Pin(const char* name);
-
-  bool getValue(void);
-  void setValue(bool);
-
-  void setMode(PinMode);
-};
-
-#endif//PIN_HPP
